backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.54.0
faiss-cpu==1.7.4
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1
pydantic==2.5.0
python-dotenv==1.0.0 